# Fix `selector-type-case` performance