--- a/.changeset/friendly-candles-deny.md
+++ b/.changeset/friendly-candles-deny.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `selector-type-case` performance