--- a/lib/rules/selector-type-case/index.js
+++ b/lib/rules/selector-type-case/index.js
@@ -22,6 +22,10 @@ const meta = {
 	fixable: true,
 };
 
+const STARTS_A_TAG_NAME_REGEX = /(?:[^.#[:\w-]|^)\w/;
+const ANY_UPPER_CASE_REGEX = /[A-Z]/;
+const ANY_LOWER_CASE_REGEX = /[a-z]/;
+
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions, context) => {
 	return (root, result) => {
@@ -48,11 +52,16 @@ const rule = (primary, secondaryOptions, context) => {
 		root.walkRules((ruleNode) => {
 			let hasComments = ruleNode.raws.selector && ruleNode.raws.selector.raw;
 			const selector = hasComments ? hasComments : ruleNode.selector;
-			const selectors = ruleNode.selectors;
 
-			if (!isStandardSyntaxRule(ruleNode)) {
-				return;
-			}
+			if (!STARTS_A_TAG_NAME_REGEX.test(selector)) return;
+
+			if (primary === 'lower' && !ANY_UPPER_CASE_REGEX.test(selector)) return;
+
+			if (primary === 'upper' && !ANY_LOWER_CASE_REGEX.test(selector)) return;
+
+			if (!isStandardSyntaxRule(ruleNode)) return;
+
+			const { selectors } = ruleNode;
 
 			if (selectors.some((s) => isKeyframeSelector(s))) {
 				return;