--- a/lib/rules/selector-type-case/index.js
+++ b/lib/rules/selector-type-case/index.js
@@ -22,7 +22,7 @@ const meta = {
 	fixable: true,
 };
 
-const STARTS_A_TAG_NAME_REGEX = /(?:[^.#[:\w-]|^)\w/;
+const STARTS_A_TAG_NAME_REGEX = /(?:[^.#[:a-zA-Z-]|^)[a-zA-Z]/;
 const ANY_UPPER_CASE_REGEX = /[A-Z]/;
 const ANY_LOWER_CASE_REGEX = /[a-z]/;