--- a/lib/rules/selector-type-no-unknown/index.js
+++ b/lib/rules/selector-type-no-unknown/index.js
@@ -24,7 +24,7 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/selector-type-no-unknown',
 };
 
-const STARTS_A_TAG_NAME_REGEX = /(?:[^.#[:\w-]|^)\w/;
+const STARTS_A_TAG_NAME_REGEX = /(?:[^.#[:a-zA-Z-]|^)[a-zA-Z]/;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions) => {