<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

Before :

```
Warnings: 0
Mean: 298.4487499090909 ms
Deviation: 21.21196081587678 ms
```

After :

```
Warnings: 0
Mean: 103.26133328571427 ms
Deviation: 9.239254247553538 ms
```